Remove sklearn dependency
For #77
