+++ src/gurobi_ml/sklearn/pipeline.py
@@ -20,7 +20,8 @@ into a :gurobipy:`model`.
 
 from ..exceptions import NoModel
 from ..modeling.basepredictor import AbstractPredictorConstr
-from .predictors_list import sklearn_predictors, sklearn_transformers, user_predictors
+from ..register_predictor import user_predictors
+from .predictors_list import sklearn_predictors, sklearn_transformers
 from .skgetter import SKgetter
